% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classification.R
\name{get_classification}
\alias{get_classification}
\title{Get taxa classification}
\usage{
get_classification(x, col = NULL, path = NULL)
}
\arguments{
\item{authority}{(character) Vector of authorities corresponding to taxonomic IDs listed
in the "authority.id" argument below. Get data.sources and IDs with
`resolve_sci_taxa`.}

\item{authority.id}{(character) Vector taxonomic IDs corresponding to an authority.}
}
\value{
(list) A list of hierarchical ranks for each taxa represented by an
    authority.id
}
\description{
Get the classification hierarchy of a taxa
}
