% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_taxa_table.R
\name{initialize_taxa_table}
\alias{initialize_taxa_table}
\title{Initialize taxa table}
\usage{
initialize_taxa_table(path, x, col)
}
\arguments{
\item{path}{A character string specifying the path to which the taxa table will be
written.}

\item{x}{A data frame containing the vector of taxa names to be cleaned.}

\item{col}{A character string specifying the column in x containing taxa names to
be cleaned.}
}
\value{
A .csv file written to path containing the fields:
    \itemize{
        \item{'taxa_raw'} Unique taxa names listed in x.
        \item{'taxa_trimmed'} The contents of taxa_raw, but with white space
        and common abbreviations (e.g. "Spp.", "C.f.") trimmed.
        \item{'taxa_replacement'} The taxa name used as a replacement for
        taxa_raw, and as implemented through `replace_taxon`.
        \item{'taxa_removed'} A logical value indicating whether the
        corresponding taxa_raw should be removed, and as implemented via
        `remove_taxon`.
        \item{'taxa_clean'} Cleaned taxa names that have been resolved to a
        taxonomic authority.
        \item{'selection'} An character string indicating which taxa_clean
        to use when more than one exists for a taxa_raw. Any character
        string in this field functions as a suitable identifier.
        \item{'difference'} Logical values indicating whether a difference
        between taxa_clean and taxa_raw exists.
        \item{'authority'} Taxonomic authorities against which taxa_clean
        was resolved.
        \item{'authority_id'} Unique identification numbers within each
        authority.
        \item{'score'} A numeric score, supplied by the authority,
        indicating the strength of match between taxa_raw and taxa_clean.
    }
}
\description{
Initialize the taxa table that will map the resolved taxa back to the
    raw taxa in the original data table, and which will be populated with
    provenance information about the taxa cleaning process.
}
