% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_taxa_authorities.R
\name{load_gna_data_sources}
\alias{load_gna_data_sources}
\title{Load and fix gna Datasources}
\usage{
load_gna_data_sources()
}
\value{
(data.frame) gna datasources from \code{taxize::gna_data_sources()}
}
\description{
Load and fix gna Datasources
}
\details{
This fixes bugs in taxize which otherwise produce inconsistent datasource names (e.g. "Integrated Taxonomic Information SystemITIS" rather than expected "ITIS")
}
\keyword{internal}
