% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_taxa_map.R
\name{read_taxa_map}
\alias{read_taxa_map}
\title{Read taxa map}
\usage{
read_taxa_map(path)
}
\arguments{
\item{path}{(character) The path to the directory containing taxa_map.csv.}
}
\value{
(data.table; data.frame) taxa_map.csv as an R object.
}
\description{
Read taxa_map.csv generated by \code{create_taxa_map()}.
}
\examples{
mypath <- dirname(system.file("taxa_map.csv", package = "taxonomyCleanr"))
taxa_map <- read_taxa_map(mypath)
taxa_map

}
