% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_taxa.R
\name{remove_taxa}
\alias{remove_taxa}
\title{Remove taxa}
\usage{
remove_taxa(x, taxa.col, input)
}
\arguments{
\item{path}{A character string specifying the path to taxa_map.csv. This table
tracks relationships between your raw and cleaned data and is operated
on by this function.}

\item{input}{A character string specifying the taxon to be removed.}
}
\value{
\itemize{
        \item{1.} An updated version of taxa_map.csv with removed taxa names.
        \item{2.} A data frame of taxa_map.csv with removed taxa names.
    }
}
\description{
Remove a taxon, and the associated record, from a data frame.
}
