% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_taxon.R
\name{remove_taxon}
\alias{remove_taxon}
\title{Remove taxon}
\usage{
remove_taxon(x, taxa.col, input)
}
\arguments{
\item{x}{The data frame containing a vector of taxa names to be operated on.}

\item{taxa.col}{A character string specifying the column name containing taxa names to
be operated on.}

\item{input}{A character string specifying the taxon to be removed.}
}
\value{
Your data frame (x) with the taxon and associated record removed.
}
\description{
Remove a taxon, and the associated record, from a data frame.
}
