% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_taxon_name.R
\name{replace_taxon_name}
\alias{replace_taxon_name}
\title{Replace taxon name}
\usage{
update_taxon_name(x, input, output)
}
\arguments{
\item{x}{The data frame containing a vector of taxa names to be operated on.}

\item{col.name}{A character string specifying the column name containing taxa names to
be operated on.}

\item{input}{A character string specifying the extant taxon name.}

\item{output}{A character string specifying the taxon name to replace the extant taxon
name.}
}
\value{
Your data frame (x) with the updated taxa names.
}
\description{
Replace a taxon name of a data frame with a user specified name.
}
