% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_names.R
\name{resolve_names}
\alias{resolve_names}
\title{Resolve taxa names}
\usage{
resolve_names(path = "", data.file = "", taxon.col = "", mode = "")
}
\arguments{
\item{path}{A path of the directory containing the data table containing taxonomic
information.}

\item{data.file}{Name of the input data table containing taxonomic data.}

\item{taxon.col}{Name of the column containing the taxonomic names including species
binomials and common names.}

\item{method}{Method for resolving your taxonomic data against an authority. There
are 2 options:
\itemize{
    \item{manual} NOT YET SUPPORTED: Outputs a table with all possible
    accepted options from which you manually select.
    \item{automatic} Output table with one-to-one matches where the resolved
    name is the first accepted match for a taxon.
    \item{interactive} You are prompted to select from a list of options
    when a one-to-one match can't be found.
}}
}
\value{
A tab delimited file in the dataset working directory titled
    \emph{taxon_map.txt} containing the relationships between your input
    taxon data and the resolved names. This file is used by
    \code{update_data} to update the taxonomic data of your data table.

    A tab delimited file in the dataset working directory titled
    \emph{taxon.txt}. This file is used to create the taxonomicCoverage
    element of the EML for your dataset.
}
\description{
Resolve input taxonomoic names against an authority and output a table
    that maps your raw taxon data to the resolved names. This map is later
    used by \code{update_data} to revise the taxon listed in your data.
}
