% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_sci_taxa.R
\name{resolve_sci_taxa}
\alias{resolve_sci_taxa}
\title{Resolve taxa with scientific names}
\usage{
resolve_sci_taxa(x = NULL, data.sources, path = NULL)
}
\arguments{
\item{x}{(character) A vector of taxa names.}

\item{data.sources}{An ordered numeric vector of ID's corresponding to data sources (i.e.
taxonomic authorities) you'd like to query, in the order of decreasing
preference. Run `view_taxa_authorities` to see data sources supported by
`resolve_sci_taxa`.}

\item{path}{A character string specifying the path to taxa_map.csv. This table
tracks relationships between your raw and cleaned data and is operated
on by this function. Create this file with `create_taxa_map`.}
}
\value{
\itemize{
        \item{1.} An updated version of taxa_map.csv containing authorities
        and corresponding IDs.
        \item{2.} A data frame of taxa_map.csv with resolved authorities and
        IDs.
    }
}
\description{
Resolve taxa to preferred authorities and get associated ID's. Run
    `view_taxa_authorities` to see supported authorities.
}
\details{
A taxa are resolved to data sources in order of listed preference. If
    an authority and ID match can not be made, then the next preferred data
    source will be queried, and so on.
}
