% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_checker.R
\name{spell_checker}
\alias{spell_checker}
\title{Check taxa spelling}
\usage{
spell_checker(x, pattern)
}
\arguments{
\item{path}{A character string specifying the path to taxa_map.csv. This table
tracks relationships between your raw and cleaned data and is operated
on by this function.}

\item{preferred.data.sources}{An ordered numeric vector of ID's corresponding to data sources (i.e.
taxonomic authorities) you'd like to query, in the order of decreasing
preference. Run `gnr_datasources()` to get valid data source options
and ID's.}
}
\value{
\itemize{
        \item{1.} An updated version of taxa_map.csv with resolved taxa
        names.
        \item{2.} A data frame of taxa_map.csv with resolved taxa names.
    }
}
\description{
Check taxa spelling against the Global Names Index (GNI) and get GNI of
    corresponding output.
}
