% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_file_connection.R
\name{test_file_connection}
\alias{test_file_connection}
\title{Test file connection}
\usage{
test_file_con(path, file.name)
}
\arguments{
\item{path}{A character string specifying the path of a file.}

\item{file.name}{A character string specifying file name.}
}
\value{
An error message if the target file is open by another application.
}
\description{
Test if file is open by another application.
}
