% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_taxa.R
\name{trim_taxa}
\alias{trim_taxa}
\title{Trim taxa}
\usage{
trim_raw_taxa_str(path = NULL, x = NULL)
}
\arguments{
\item{path}{A character string specifying the path to taxa_map.csv. This table
tracks relationships between your raw and cleaned data and is operated
on by this function.}

\item{x}{(character) A vector of taxa to be trimmed.}
}
\value{
\itemize{
        \item{1.} An updated version of taxa_map.csv with trimmed taxa strings.
        \item{2.} A data frame of taxa_map.csv with trimmed taxa strings.
    }
}
\description{
Trim excess text from taxa names. Doing this before querying a taxonomic
    authority reduces the frequency of mismatches.
}
\details{
List of conditions `trim_taxa` addresses:
    \itemize{
        \item{'White spaces, trailing and leading'} White spaces are common.
        \item{'Species abbreviations trailing a genus value'} E.g. "Sp.",
        "Spp.", "Cf.", "Cf...", etc.
    }
}
