% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_taxon.R
\name{trim_taxon}
\alias{trim_taxon}
\title{Trim taxon strings}
\usage{
trim_raw_taxon_str(path)
}
\arguments{
\item{path}{A character string specifying the path to taxa_map.csv. This table
tracks relationships between your raw and cleaned data and is operated
on by this function.}
}
\value{
\itemize{
        \item{1.} An updated version of taxa_map.csv with trimmed taxa strings.
        \item{2.} A data frame of taxa_map.csv with trimmed taxa strings.
    }
}
\description{
Trim excess text from taxa names. Doing this before querying a taxonomic
    authority reduces the frequency of mismatches.
}
\details{
List of conditions `trim_taxon` addresses:
    \itemize{
        \item{'White spaces, trailing and leading'} White spaces are common.
        \item{'Species abbreviations trailing a genus value'} E.g. "Sp.",
        "Spp.", "Cf.", "Cf...", etc.
    }
}
