% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Update data}
\usage{
update_data(path = "", data.file = "", taxon.col = "")
}
\arguments{
\item{path}{A path of the directory containing the data table containing taxonomic
information and the file \emph{taxon_map.txt} for this dataset.}

\item{data.file}{Name of the data table containing taxonomic data.}

\item{taxon.col}{Name of the column containing the taxonomic names including species
binomials and common names.}
}
\value{
A copy of your data table containing revised taxonomic data (for those
    taxa that could be resolved, non-resolved taxa are not updated). The
    copy of your data is appended with \emph{taxon_cleaned_} and the current
    date-time stamp. This copy is made in the directory specified by path.
}
\description{
Update the taxonomic information in your data table based on the
    mapping containted in \emph{taxon_map.txt} created by the \code{resolve_names}
    function.
}
