% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arguments.R
\name{validate_arguments}
\alias{validate_arguments}
\title{Validate arguments of taxonomyCleanr functions}
\usage{
validate_arguments(fun.name, fun.args)
}
\arguments{
\item{fun.name}{(character) Function name}

\item{fun.args}{(named list) Function arguments and values passed from the calling
function.}
}
\description{
Validate input arguments to taxonomyCleanr functions.
}
\keyword{internal}
