% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_taxa_authorities.R
\name{view_taxa_authorities}
\alias{view_taxa_authorities}
\title{View taxonomic authorities}
\usage{
view_taxa_authorities()
}
\value{
A data frame and view of taxonomic authorities and corresponding
    identifiers supported by `resolve_taxa` and `resolve_common`.
}
\description{
List taxonomic authorities supported by the `taxonomyCleanr`.
}
\details{
View taxonomic authorities supported by `resolve_taxa` and
    `resolve_common`.
}
