% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_unique.R
\name{view_unique}
\alias{view_unique}
\title{View unique occurence and counts}
\usage{
view_unique(x, col)
}
\arguments{
\item{x}{A data frame containing the vector of taxa names to be cleaned.}

\item{col}{A character string specifying the column in x containing taxa names to
be cleaned.}
}
\value{
A data frame with taxa and associated counts found in the input, sorted
    alphabetically by taxa.
}
\description{
View unique occurences and counts of taxa to help identify misspelled
    taxon.
}
