% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_taxa_map.R
\name{write_taxa_map}
\alias{write_taxa_map}
\title{Write taxa map}
\usage{
write_taxa_map(x, path)
}
\arguments{
\item{x}{A data frame representation of taxa_map.csv. Data frame contents will
be written to file.}

\item{path}{A character string specifying the path of where taxa_map.csv
should be written or is written.}
}
\value{
The taxa_map.csv file, or an updated version of this file to path.
}
\description{
Writes information created in the `taxonomyCleanr` workflow to
    taxa_map.csv. An error is issued if the file is open by another
    application.
}
\seealso{
initialize_taxa_table
}
\keyword{internal}
