% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_taxa.R
\name{count_taxa}
\alias{count_taxa}
\title{Get counts of unique taxa}
\usage{
count_taxa(x, col = NULL, path = NULL)
}
\arguments{
\item{x}{(data frame or character) A data frame or vector of taxa names to be
cleaned.}

\item{col}{(character) A character string specifying the column in x containing
taxa names to be cleaned. NOTE: Don't use this argument if x is of
character class.}

\item{path}{(character) A character string specifying the path to taxa_map.csv.
This table tracks relationships between your raw and cleaned data and
is used by this function.}
}
\value{
(data frame) A data frame with taxa and associated counts found in the
    input, sorted alphabetically by taxa.
}
\description{
View unique occurences and counts of taxa to help identify spelling
    errors.
}
