% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classification.R
\name{get_classification}
\alias{get_classification}
\title{Get taxa classification hierarchy}
\usage{
get_classification(taxa.clean, authority, authority.id, path = NULL)
}
\arguments{
\item{taxa.clean}{(character) Name of taxa resolved to an authority.}

\item{authority}{(character) Vector of authorities corresponding to taxonomic IDs listed
in the "authority.id" argument below. Get data.sources and IDs with
`resolve_sci_taxa`.}

\item{authority.id}{(character) Vector taxonomic IDs corresponding to an authority.}

\item{path}{A character string specifying the path to taxa_map.csv. This table
tracks relationships between your raw and cleaned data and is operated
on by this function. Create this file with `create_taxa_map`.}
}
\value{
(list) A list taxa with corresponding hierarchical names and ranks.
}
\description{
Get the classification hierarchy of taxa.
}
