% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_taxonomicCoverage.R
\name{make_taxonomicCoverage}
\alias{make_taxonomicCoverage}
\title{Make taxonomicCoverage EML}
\usage{
make_taxonomicCoverage(taxa.clean, authority, authority.id, path = NULL)
}
\arguments{
\item{taxa.clean}{(character) Taxa names that have been resolved to a taxonomic authority.}

\item{authority}{(character) Authorities to which the taxa have been resolved. Valid
inputs are created by `resolve_sci_taxa`.}

\item{authority.id}{(character) Taxonomic IDs corresponding the authority to which taxa
have been resolved.}

\item{path}{(character) The path to which the taxonomicCoverage will be written, and
or to where taxa_map.csv is located.}

\item{write.file}{(logical) Flag to indicate if the resulting EML taxonomicCoverage element
should be written to file as an xml document.}
}
\value{
The taxonomicCoverage EML as an 'emld' 'list' object, and or written to
    the file taxonomicCoverage.xml in the directory specified by path.
}
\description{
Make the hierarchical taxonomicCoverage EML node set.
}
