% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_taxonomicCoverage.R
\name{make_taxonomicCoverage}
\alias{make_taxonomicCoverage}
\title{Make taxonomicCoverage EML}
\usage{
make_taxonomicCoverage(
      taxa.clean,
      authority,
      authority.id,
      path,
      write.file = TRUE
    )
}
\arguments{
\item{taxa.clean}{(character) Taxa names as they appear in your dataset, or as they appear
in an authority system. Best practice is to align the names in your
dataset with those of the authority system.}

\item{authority}{(character) Authority to which the taxa have been resolved. Valid inputs
are created by \code{resolve_sci_taxa()}.}

\item{authority.id}{(character) Authority specific identifiers for the resolved taxa.}

\item{path}{(character) Path to where taxonomicCoverage.xml will be written, and or
to where taxa_map.csv is located.}

\item{write.file}{(logical) Whether taxonomicCoverage.xml should be written to file.
Default is \code{TRUE}.}
}
\value{
The taxonomicCoverage as an 'emld' 'list' object (required for use in
    the \code{EML} library), and written to taxonomicCoverage.xml if
    \code{write.file = TRUE}.
}
\description{
Create the hierarchical rank specific EML for taxa that have been
    resolved to an authority system (e.g. ITIS).
}
\examples{
# Create taxonomicCoverage from taxa_map.csv ----------------------------------

# Copy taxa_map.csv from the taxonomyCleanr library to tempdir()
file.copy(
  from = system.file('/taxa_map_resolve_sci_taxa/taxa_map.csv', package = 'taxonomyCleanr'),
  to = tempdir()
)

# Create taxonomicCoverage
output <- make_taxonomicCoverage(
  path = tempdir(),
)

# View taxonomicCoverage list object
output

# Verify taxonomicCoverage.xml has been written to file
file.exists(paste0(tempdir(), '/taxonomicCoverage.xml'))

# Create taxonomicCoverage from argument inputs -------------------------------

# Create taxonomicCoverage
output <- make_taxonomicCoverage(
  taxa.clean = c('Crepis tectorum', 'Euphorbia glyptosperma'),
  authority = c('ITIS', 'ITIS'),
  authority.id = c('37212', '28074'),
  path = tempdir(),
)

# View taxonomicCoverage list object
output

# Verify taxonomicCoverage.xml has been written to file
file.exists(paste0(tempdir(), '/taxonomicCoverage.xml'))

}
