% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_taxa.R
\name{remove_taxa}
\alias{remove_taxa}
\title{Remove taxa}
\usage{
remove_taxa(input, x = NULL, col = NULL, path = NULL)
}
\arguments{
\item{input}{(character) A character string specifying the taxon to be removed.}

\item{x}{(character vector or data frame) Data containing taxa to be removed.}

\item{col}{(character) A character string specifying the column in x containing
taxa names to be cleaned. NOTE: Don't use this argument if x is a
vector of character strings.}

\item{path}{(character) A character string specifying the path to taxa_map.csv.}
}
\value{
\itemize{
        \item{1.} An updated version of taxa_map.csv with removed taxa names.
        \item{2.} A data frame of taxa_map.csv with removed taxa names, if
        the path argument is supplied.
    }
}
\description{
Remove a taxon, and the associated record, from a data frame.
}
