% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_taxa.R
\name{replace_taxa}
\alias{replace_taxa}
\title{Replace taxa}
\usage{
replace_taxa(path, input, output)
}
\arguments{
\item{path}{A character string specifying the path to taxa_map.csv. This table
tracks relationships between your raw and cleaned data and is operated
on by this function.}

\item{input}{A character string specifying an existing taxon name.}

\item{output}{A character string specifying a replacement taxon name.}
}
\value{
\itemize{
        \item{1.} An updated version of taxa_map.csv with new taxa names.
        \item{2.} A data frame of taxa_map.csv with new taxa names.
    }
}
\description{
Replace a taxon name with another.
}
