% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_common.R
\name{resolve_common}
\alias{resolve_common}
\title{Resolve common}
\usage{
resolve_common(path, data.sources)
}
\arguments{
\item{path}{A character string specifying the path to taxa_map.csv. This table
tracks relationships between your raw and cleaned data and is operated
on by this function. Create this file with `initialize_taxa_map`.}

\item{data.sources}{An ordered numeric vector of ID's corresponding to data sources (i.e.
taxonomic authorities) you'd like to query, in the order of decreasing
preference. Run `view_authorities` to get valid data source options
and ID's.}
}
\value{
\itemize{
        \item{1.} An updated version of taxa_map.csv containing authorities
        and corresponding IDs.
        \item{2.} A data frame of taxa_map.csv with resolved authorities and
        IDs.
    }
}
\description{
Resolve common names to preferred authorities and get associated ID's.
}
\details{
Common names are resolved to data sources in order of listed preference.
    If an authority and ID match can not be made, then the next preferred data
    source will be queried, and so on.
}
