% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_checker.R
\name{spell_checker}
\alias{spell_checker}
\title{Check taxa spelling}
\usage{
spell_checker(x, pattern)
}
\arguments{
\item{x}{A character string, or vector of character strings, representing taxa
names.}

\item{pattern}{A character string, or vector of character strings, representing the
taxa pattern to be searched.}

\item{pds}{A vector of character strings representing authorities to query as specified in GNI
resolver}
}
\value{
A data frame with the search term, returned value, GNI identification
    number
}
\description{
Check taxa spelling against the Global Names Index (GNI) and get GNI of
    corresponding output.
}
