% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_checker.R
\name{spell_checker}
\alias{spell_checker}
\title{Check taxa spelling}
\usage{
spell_checker(x, pattern)
}
\arguments{
\item{x}{A character string, or vector of character strings, representing taxa
names.}

\item{preferred.data.sources}{An ordered numeric vector of ID's corresponding to data sources (i.e.
taxonomic authorities) you'd like to query, in the order of decreasing
preference. Run `gnr_datasources()` to get valid data source options
and ID's.}
}
\value{
A data frame with the input taxon (search_term), corresponding match
    from the Global Names Resolver (result), an indicator of whether
    search_term and result differ (difference), the data source resolved
    against (source), the data source ID for the resolved taxon (id), and
    the match score (score).
}
\description{
Check taxa spelling against the Global Names Index (GNI) and get GNI of
    corresponding output.
}
