% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_taxa.R
\name{trim_taxa}
\alias{trim_taxa}
\title{Trim taxa}
\usage{
trim_taxa(x = NULL, path = NULL)
}
\arguments{
\item{x}{(character) A list of taxa names}

\item{path}{(character) The path to the directory containing taxa_map.csv}
}
\value{
\itemize{
        \item{If using \code{x}, then an updated list of taxa is returned.}
        \item{If using \code{path}, then an updated version of taxa_map.csv
        is written to file and returned as a data frame.}
    }
}
\description{
Trim excess text from taxa names. Doing this before querying a taxonomic
    authority improves matching.
}
\details{
Characters trimmed by this function:
    \itemize{
        \item{White spaces (trailing and leading)}
        \item{Species abbreviations following a genus value (e.g. "Sp.",
        "Spp.", "Cf.", "Cf...", etc.)}
    }
}
\examples{
# Input a list of names

my_taxa <- c("Koeleria_cristata", "Cyperus sp.", "Poa Cf.   ")
my_taxa <- trim_taxa(x = my_taxa)
my_taxa

# Input taxa_map.csv

data <- data.table::fread(system.file("example_data.txt", package = 'taxonomyCleanr'))
my_path <- tempdir()
tm <- create_taxa_map(path = my_path, x = data, col = "Species")
tm <- trim_taxa(path = my_path)
tm

}
