% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_taxa_str.R
\name{trim_taxa_str}
\alias{trim_taxa_str}
\title{Trim taxa strings}
\usage{
trim_raw_taxon_str(x)
}
\arguments{
\item{x}{A character string, or vector of character strings, representing taxa
names to be trimmed.}
}
\value{
A character string, or vector of strings, with excess characters
    trimmed. See details for trimming conditions.
}
\description{
Trim excess text from raw taxa strings. Doing this before querying an
    authority for a match may reduce the frequency of mismatches.
}
\details{
List of conditions `trim_taxa_str` addresses:
    \itemize{
        \item{'White spaces, trailing and leading'} White spaces are common.
        \item{'Species abbreviations trailing a genus value'} E.g. "Sp.",
        "Spp.", "Cf.", "Cf...", etc.
    }
}
