% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Update data}
\usage{
update_data(path, x, col, retain.raw, sep)
}
\arguments{
\item{path}{A path of the directory containing taxa_map.csv and the raw data table.}

\item{x}{A data frame containing the vector of taxa names to be updated.}

\item{col}{A character string specifying the column in x containing taxa names to
be updated.}

\item{sep}{The column delimiter to use when writting the table to file. Can be ","
or "\t".}
}
\value{
A copy of your data table (taxonomyCleanr_output) containing appended
    columns containing:
    \itemize{
        \item{taxa_clean} Both resolved and unresolved taxa.
        \item{taxa_rank} Taxonomic rank of resolved taxa, otherwise NA.
        \item{taxa_authority} Authority system of resolved taxa, otherwise
        NA.
        \item{taxa_authority_id} Resolved taxa identifiers from
        corresponding authority system.
    }
    Unresolvable taxa will have NA values for the appended columns.
}
\description{
Update the taxonomic information in your data table based on the
    mapping containted in \emph{taxa_map.csv}.
    function.
}
