% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Update data}
\usage{
update_data(path = "", data.file = "", col = "")
}
\arguments{
\item{path}{A path of the directory containing taxa_map.csv and the raw data table.}

\item{x}{A data frame containing the vector of taxa names to be updated.}

\item{col}{A character string specifying the column in x containing taxa names to
be updated.}
}
\value{
A copy of your data table containing appended columns containing:
    \itemize{
        \item{resolved_taxa} Resolved taxa names.
        \item{rank} Taxonomic rank.
        \item{authority} Authority system the resolved names correspond with.
        \item{authority_id} Taxonomic IDs of the authority system.

    }
}
\description{
Update the taxonomic information in your data table based on the
    mapping containted in \emph{taxa_map.csv}.
    function.
}
