% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_authorities.R
\name{view_authorities}
\alias{view_authorities}
\title{View authorities}
\usage{
view_authorities()
}
\value{
A data frame and view of taxonomic authorities and corresponding
    identifiers supported by `get_authorities`.
}
\description{
List taxonomic authorities supported by the `taxonomyCleanr`.
}
\details{
This is a wrapper function to `taxize::gnr_datasources`.
}
