% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_unique.R
\name{view_unique}
\alias{view_unique}
\title{View unique occurence and counts}
\usage{
view_unique(x)
}
\arguments{
\item{x}{A character string, or vector of character strings, representing taxa
names to be analyzed.}
}
\value{
A data frame with taxa and associated counts found in the input, sorted
    alphabetically by taxa.
}
\description{
View unique occurences and counts of taxa to help identify misspelled
    taxon.
}
